ai.services/photo-crawler/requirements.txt
playwright>=1.40.0
playwright-stealth>=0.1.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
reportlab>=4.0.7
Pillow>=10.1.0
